matchBuild10 updates

+++ matchBuild10/matchBuild10/CardMatchingGame.h
@@ -15,11 +15,10 @@
 
 -(instancetype)initWithCardCount:(NSUInteger)count usingDeck:(Deck *)deck;
 
-
-
 -(void)chooseCardAtIndex:(NSUInteger)index;
 -(Card *)cardAtIndex:(NSUInteger)index;
 
-@property (nonatomic,readonly)NSInteger score;
+@property (nonatomic,readonly)int score;
+
 
 @end
